fix: Unaligned pointers due to 32-bit Windows ABI
The Windows ABI for structs which contain 8-byte aligned types do not
always force the structure alignment to 8-bytes, which can cause an
underalignment of the structure pointers when returned from the Win32
API functions, resulting in undefined behavior. This UB is unlikely to
ever be seen in practice, since x86 allows unaligned reads just fine and
no other architecture supports 32-bit Windows, but it does trigger a
debug assertion on newer nightlies.